pyproject.toml
[tool.poetry]
name = "backend-homework"
version = "0.1.0"
description = "Center of Excellence backend home assignment"
authors = ["CoE"]

[tool.poetry.dependencies]
python = "^3.8"
kafka-python = "^2.0.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
